requirements.txt
numpy
pandas
scikit-learn
xgboost
joblib
pytest
pyyaml
yfinance

